Fix a bunch of compiler warnings. Also fix a peculiar bug in WriteBSP where EmitAreaPortals was called for each BSP submodel. Now we only call it once, in EndBSPFile.

+++ src/tools/q2wmap/main.c
@@ -506,8 +506,8 @@ int32_t main(int32_t argc, char **argv) {
 	const time_t duration = end - start;
 	Com_Print("\nTotal Time: ");
 	if (duration > 59)
-		Com_Print("%ld Minutes ", duration / 60);
-	Com_Print("%ld Seconds\n", duration % 60);
+		Com_Print("%d Minutes ", (int32_t) (duration / 60));
+	Com_Print("%d Seconds\n", (int32_t) (duration % 60));
 
 	Com_Shutdown(NULL);
 }
